requirements.txt
# Web Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1

# Database
psycopg2-binary==2.9.9
alembic==1.13.1

# API Requests
requests==2.31.0

# Azure Integration
msal==1.26.0
azure-storage-blob==12.19.0

# Email for alerts
Flask-Mail==0.9.1

# Environment variables
python-dotenv==1.0.0

# Scheduler for alerts
APScheduler==3.10.4

# Web server for production
gunicorn==21.2.0

# CORS support
Flask-CORS==4.0.0

# Date/Time handling
python-dateutil==2.8.2
